Changed model Parameter and parsing in DialogWindow.

+++ ParameterReferenceBook/Models/Parameter.cs
@@ -5,8 +5,8 @@ namespace ParameterReferenceBook
         public long IdParameter { get; set; }
         public string Name { get; set; }
         public long IdTypeParameter { get; set; }
-        public short? MinValue { get; set; }
-        public short MaxValue { get; set; }
+        public float? MinValue { get; set; }
+        public float MaxValue { get; set; }
         public string Description { get; set; }
 
         public TypeParameter TypeParameter { get; set; }
